requirements.txt
cvlib==0.2.7
ImageHash==4.3.1
PyQt5==5.15.7
pytest==7.2.0
tqdm==4.66.2
opencv-python-headless<=4.8.0.0
sahi==0.11.15
torch==2.2.2
inference-gpu[yolo-world]

